sql-agent/requirements.txt
langchain==0.1.4
langchain-core==0.1.14
langchain-community==0.0.13
langchain-mistralai==0.0.3
langgraph==0.0.11
snowflake-connector-python<3.6.0
snowflake-snowpark-python<1.11.0
pydantic>=2.0.0,<3.0.0
python-dotenv==1.0.0
typing-extensions>=4.5.0
requests==2.31.0